municipalities/requirements.txt
affine==2.3.0
asgiref==3.3.4
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
contextily==1.1.0
cycler==0.10.0
decorator==4.4.2
descartes==1.1.0
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-gis==0.17
djangorestframework-simplejwt==4.6.0
Fiona==1.8.19
geographiclib==1.50
geopandas==0.9.0
geoplot==0.4.1
geopy==2.1.0
geos==0.2.3
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
mapclassify==2.4.2
matplotlib==3.4.1
mercantile==1.2.1
munch==2.5.0
networkx==2.5.1
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
psycopg2==2.8.6
PyJWT==2.0.1
pyparsing==2.4.7
pyproj==3.0.1
pyshp==2.1.3
python-dateutil==2.8.1
pytz==2021.1
rasterio==1.2.2
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
six==1.15.0
snuggs==1.4.7
sqlparse==0.4.1
threadpoolctl==2.1.0
urllib3==1.26.4
cartopy==0.17.0
gunicorn==20.1.0
